+++ cloud-project/src/main/java/it/unipi/hadoop/LetterCountReducer.java
@@ -2,16 +2,16 @@ package it.unipi.hadoop;
 import org.apache.hadoop.io.*;
 import org.apache.hadoop.mapreduce.Reducer;
 import java.io.IOException;
-import org.apache.hadoop.io.IntWritable;
+import org.apache.hadoop.io.LongWritable;
 
-public class LetterCountReducer extends Reducer<Text, IntWritable, Text, IntWritable> {
+public class LetterCountReducer extends Reducer<Text, LongWritable, Text, LongWritable> {
 
-    private IntWritable result = new IntWritable();
+    private LongWritable result = new LongWritable();
    
     @Override
-    public void reduce(Text key, Iterable<IntWritable> values, Context context) throws IOException, InterruptedException {
-        int sum = 0;
-        for (IntWritable val : values) {
+    public void reduce(Text key, Iterable<LongWritable> values, Context context) throws IOException, InterruptedException {
+        long sum = 0;
+        for (LongWritable val : values) {
             sum += val.get();
         }
         result.set(sum);
